various fixes, action dropdown working correctly now
