+++ app/src/main/res/layout/activity_request.xml
@@ -21,6 +21,7 @@
 
 		<com.google.android.material.floatingactionbutton.ExtendedFloatingActionButton
 			android:id="@+id/fab"
+			style="@style/StaxFabExtended"
 			android:layout_width="wrap_content"
 			android:layout_height="wrap_content"
 			android:layout_marginHorizontal="@dimen/margin_13"
