+++ app/src/main/res/layout/transfer_card_edit.xml
@@ -18,7 +18,8 @@
 	<com.hover.stax.actions.ActionSelect
 		android:id="@+id/action_select"
 		android:layout_width="match_parent"
-		android:layout_height="wrap_content" />
+		android:layout_height="wrap_content"
+		android:visibility="gone"/>
 
 	<include  layout="@layout/recipient_input" />
