+++ app/src/main/res/values/styles.xml
@@ -83,6 +83,21 @@
 		<item name="textAllCaps">false</item>
 	</style>
 
+	<style name="StaxFabExtended" parent="Widget.MaterialComponents.ExtendedFloatingActionButton">
+		<item name="android:padding">@dimen/margin_13</item>
+		<item name="android:layout_marginBottom">@dimen/margin_89</item>
+		<item name="android:radius">5dp</item>
+		<item name="android:textColor">@color/colorPrimaryDark</item>
+		<item name="backgroundTint">@color/brightBlue</item>
+		<item name="elevation">1dp</item>
+		<item name="android:textSize">@dimen/text_16</item>
+		<item name="android:textAllCaps">false</item>
+		<item name="textAllCaps">false</item>
+		<item name="android:fontFamily">font-medium</item>
+		<item name="fontFamily">font-medium</item>
+		<item name="shapeAppearanceOverlay">@style/ShapeAppearanceOverlayExtended</item>
+	</style>
+
 	<style name="HoverButton" parent="StaxButton">
 		<item name="android:paddingStart">@dimen/margin_13</item>
 		<item name="android:paddingEnd">@dimen/margin_13</item>
